# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "منصة التعليم - API",
    "description": "واجهة عربية أولًا لإدارة التعليم والعمليات المالية",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/financial-transactions": {
      "get": {
        "tags": [
          "FinancialTransactions"
        ],
        "parameters": [
          {
            "name": "financialPeriodId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/operating-expenses/{id}/{decision}": {
      "post": {
        "tags": [
          "OperatingExpenses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "decision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/student-payments": {
      "get": {
        "tags": [
          "StudentPayments"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "subjectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateStudentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateStudentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateStudentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.StudentResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teachers": {
      "get": {
        "tags": [
          "Teachers"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "curriculumId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.PageResult`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Teachers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateTeacherRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateTeacherRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Application.CreateTeacherRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationPlatform.Application.ApiResponse`1[[EducationPlatform.Application.TeacherResponse, EducationPlatform.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/users/staff": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Api.Controllers.UsersController.StaffRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Api.Controllers.UsersController.StaffRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EducationPlatform.Api.Controllers.UsersController.StaffRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/users/{id}/activate": {
      "post": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/users/{id}/deactivate": {
      "post": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EducationPlatform.Api.Controllers.UsersController.StaffRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "contactInformation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EducationPlatform.Application.CreateStudentRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "parentName": {
            "type": "string",
            "nullable": true
          },
          "parentPhoneNumber": {
            "type": "string",
            "nullable": true
          },
          "gradeLevelId": {
            "type": "string",
            "format": "uuid"
          },
          "curriculumId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EducationPlatform.Application.CreateTeacherRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "whatsApp": {
            "type": "string",
            "nullable": true
          },
          "defaultPerSessionRate": {
            "type": "number",
            "format": "double"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "curriculumIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "preferredPayoutMethod": {
            "type": "string",
            "nullable": true
          },
          "eWalletNumber": {
            "type": "string",
            "nullable": true
          },
          "instaPayIdentifier": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "أدخل JWT الناتج من /api/v1/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}